paas-metrics: move logit client to common-go/basic_logit_client for reuse
