+++ common-go/basic_logit_client/go.mod
@@ -1,0 +1,5 @@
+module github.com/alphagov/paas-cf/common-go/basic_logit_client
+
+go 1.18
+
+require code.cloudfoundry.org/lager v2.0.0+incompatible // indirect
