+++ tools/metrics/gauges_billing_collector_performance.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/alphagov/paas-cf/tools/metrics/pkg/logit"
+	logit "github.com/alphagov/paas-cf/common-go/basic_logit_client"
 	m "github.com/alphagov/paas-cf/tools/metrics/pkg/metrics"
 )
